requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
elevenlabs==0.2.26
pysrt==1.1.2
pathlib 